--- packages/blockchain-wallet-v4/src/redux/addressLabels/sagaRegister.js
@@ -1,12 +1,0 @@
-import { takeLatest } from 'redux-saga/effects'
-import * as AT from '../actionTypes'
-import sagas from './sagas'
-
-export default () => {
-  const addressLabelSagas = sagas()
-
-  return function * () {
-    yield takeLatest(AT.wallet.REFRESH_WRAPPER, addressLabelSagas.generateAddressLabels)
-    yield takeLatest(AT.wallet.SET_WRAPPER, addressLabelSagas.generateAddressLabels)
-  }
-}
