+++ packages/blockchain-wallet-v4/src/redux/rootSaga.js
@@ -1,5 +1,4 @@
 import { all, fork } from 'redux-saga/effects'
-import addressLabels from './addressLabels/sagaRegister'
 import data from './data/sagaRegister'
 import kvStore from './kvStore/sagaRegister'
 import walletOptions from './walletOptions/sagaRegister'
@@ -9,7 +8,6 @@ import wallet from './wallet/sagaRegister'
 
 export default ({ api, options }) => function * () {
   yield all([
-    fork(addressLabels()),
     fork(data({ api, options })),
     fork(kvStore({ api })),
     fork(walletOptions({ api, options })),
