+++ packages/blockchain-wallet-v4/src/redux/wallet/sagaRegister.js
@@ -10,5 +10,7 @@ export default ({ api }) => {
     yield takeLatest(AT.SET_DEFAULT_ACCOUNT, walletSagas.refetchContextData)
     yield takeLatest(AT.SET_ADDRESS_ARCHIVED, walletSagas.refetchContextData)
     yield takeLatest(AT.SET_ACCOUNT_ARCHIVED, walletSagas.refetchContextData)
+    yield takeLatest(AT.REFRESH_WRAPPER, walletSagas.generateAddressLabels)
+    yield takeLatest(AT.SET_WRAPPER, walletSagas.generateAddressLabels)
   }
 }
