+++ packages/blockchain-wallet-v4/src/redux/wallet/sagas.js
@@ -1,4 +1,5 @@
-import { call, put, select } from 'redux-saga/effects'
+import { call, put, select, fork } from 'redux-saga/effects'
+import { delay } from 'redux-saga'
 import BIP39 from 'bip39'
 import Bitcoin from 'bitcoinjs-lib'
 import { prop, compose, endsWith, repeat, range, map, propSatisfies,
@@ -8,8 +9,9 @@ import Task from 'data.task'
 import * as A from './actions'
 import * as S from './selectors'
 import { fetchData } from '../data/bitcoin/actions'
+import * as actions from '../actions'
 
-import { Wrapper, Wallet } from '../../types'
+import { Wrapper, Wallet, HDAccount } from '../../types'
 import { generateMnemonic } from '../../walletCrypto'
 
 const taskToPromise = t => new Promise((resolve, reject) => t.fork(reject, resolve))
@@ -139,6 +141,28 @@ export default ({ api }) => {
     yield put(fetchData())
   }
 
+  const generateAddressLabels = function * ({payload}) {
+    yield fork(generateAddressesNonBlocking, payload)
+  }
+
+  const generateAddressesNonBlocking = function * (payload) {
+    yield call(delay, 4000)
+    const accounts = Wallet.selectHDAccounts(payload.wallet)
+
+    let addressLabels = accounts.map((account) => {
+      const hd = accounts.get(account.index)
+      return account.address_labels.map((label) => ({
+        address: HDAccount.getReceiveAddress(hd, label.index),
+        label: label.label,
+        index: label.index
+      }))
+    }).flatten().toArray()
+
+    for (let i in addressLabels) {
+      yield put(actions.data.bitcoin.addAddressLabel(addressLabels[i].address, addressLabels[i].label))
+    }
+  }
+
   return {
     toggleSecondPassword,
     createWalletSaga,
@@ -151,6 +175,7 @@ export default ({ api }) => {
     upgradeToHd,
     resetWallet2fa,
     refetchContextData,
-    resendSmsLoginCode
+    resendSmsLoginCode,
+    generateAddressLabels
   }
 }
